Player Animation
Made the player class with idle. movement, and attack animation.
Also made the movement more arpg like with right click moving and left click attacking.
Im going to add sound effects to the movement and the slash to hit one of the required benchmarks.
I will also be changing how the Main camera works and adding a pause function.